minor modifications in detecting/classifying the input data as well as processing the short frames after histogramming

+++ niftypet/nipet/lm/mmrhist.py
@@ -154,8 +154,8 @@ def hist(datain, txLUT, axLUT, Cnt, frms=np.array([0], dtype=np.uint16), use_sto
         return
 
     #short (interval) projection views
-    pvs_sgtl = np.array( hstout['pvs']>>8, dtype=float)
-    pvs_crnl = np.array( np.bitwise_and(hstout['pvs'], 255), dtype=float )
+    pvs_sgtl = np.array( hstout['pvs']>>8, dtype=np.float32)
+    pvs_crnl = np.array( np.bitwise_and(hstout['pvs'], 255), dtype=np.float32)
 
     cmass = Cnt['SO_VXZ']*ndi.filters.gaussian_filter(hstout['mss'], cmass_sig, mode='mirror')
     if Cnt['VERBOSE']: print 'i> centre of mass of axial radiodistribution (filtered with Gaussian of SD =', cmass_sig, '):  COMPLETED.'
